Update FireProjectileScript.cs
Addressed from PR:
- Unnecessary Debug
- Orphaned semi-colon
- Hit cannot be null (TIL!)

Not addressed:
- Trying again after some time

I did not address the issue because this script is invoked once the spawn has already happened - in which case the game already thinks the gears are in motion with the new beacon. Also was not sure what a sane waiting period would be - so for now sticking with "get fired straight down" since it happens to work in the Alpha level.