Improve landing animation
Changed from a roll landing animation to a hard landing animation. It
looks better and it shouldn't suffer from the reported circular motion
issue.

I believe the circular motion issue (i.e. the player doesn't roll in a
particular direction, they just kind of roll in a circle?) is caused by
an interaction between the rolling animation and the uneven terrain.
Not sure how to fix it without having to tweak/change the rolling
animation, and making it look unnatural.